Move Checkpoint into its own module.
